requirements.txt
langchain
openai
huggingface-hub
python-dotenv
streamlit
